refactor: moved aws storage interface files to shared
