+++ server/tests/unit/canvas.unit.test.ts
@@ -30,7 +30,7 @@ mock.module('crypto', () => ({
     randomUUID: () => 'fixed-uuid',
 }))
 
-mock.module('../../src/modules/project/project-storage', () => ({
+mock.module('../../src/shared/project-storage', () => ({
     assetKey: (projectId: string, assetPath: string) =>
         `projects/${projectId}/assets/${assetPath.replace(/^\/+/, '')}`,
     assetPrefix: (projectId: string) => `projects/${projectId}/assets/`,
